refactor: move format and holidays utils to lib
  - Moved format and holidays folders to lib and updated all the exports
and imports of the funcs included there
