+++ src/lib/format/formatDateInput.ts
